Guardar datos en base de datos

+++ SmartpotModel/Sensor_Temperature.cpp
@@ -8,7 +8,7 @@
 /**
  * Sensor_Temperature implementation
  */
-System::Sensor_Temperature::Sensor_Temperature(float value)
+System::Sensor_Temperature::Sensor_Temperature(int value)
 {
 	//throw gcnew System::NotImplementedException();
 	this->Value = value;
