+++ SmartpotModel/Sensor_Uv.cpp
@@ -9,7 +9,7 @@
  * Sensor_Uv implementation
  */
 
-System::Sensor_Uv::Sensor_Uv(float value)
+System::Sensor_Uv::Sensor_Uv(int value)
 {
 	//throw gcnew System::NotImplementedException();
 	this->Value = value;
