+++ SmartpotModel/Sensor_Uv.h
@@ -10,10 +10,10 @@ namespace System {
     public:
         property int Id;
         property bool Connected;
-        property float Value;
+        property int Value;
 
         Sensor_Uv() {}
-        Sensor_Uv(float value);
+        Sensor_Uv(int value);
     };
 }
