+++ SmartpotModel/Sensor_humidity.cpp
@@ -8,7 +8,7 @@
 /**
  * Sensor_humidity implementation
  */
-System::Sensor_humidity::Sensor_humidity(float value)
+System::Sensor_humidity::Sensor_humidity(int value)
 {
 	//throw gcnew System::NotImplementedException();
 	this->Value = value;
